submit D.cpp to 'D - すぬけ君の塗り絵 / Snuke's Coloring' (abc045) [C++14 (GCC 5.4.1)]

+++ 2017/1106_ABC045/D.cpp
@@ -33,7 +33,7 @@ typedef long long ll;
 // const int C = 1e6+10;
 // const ll M = 1000000007;
 
-int H, W, N;
+ll H, W, N;
 typedef tuple<int, int> point;
 set<point> S;
